Move constants to utils
